+++ packages/amplify-e2e-tests/src/utils/sdk-calls.ts
@@ -1,5 +1,16 @@
 import * as AWS from 'aws-sdk';
 
+
+const getDDBTable = async (tableName: string, region: string) => {
+  const service = new AWS.DynamoDB({ region });
+  return await service.describeTable({ TableName: tableName }).promise();
+};
+
+const checkIfBucketExists = async (bucketName: string, region: string) => {
+  const service = new AWS.S3({ region });
+  return await service.headBucket({ Bucket: bucketName }).promise();
+};
+
 const getUserPool = async (userpoolId, region) => {
   AWS.config.update({ region });
   const CognitoIdentityServiceProvider = AWS.CognitoIdentityServiceProvider;
@@ -7,6 +18,20 @@ const getUserPool = async (userpoolId, region) => {
   try {
     res = await new CognitoIdentityServiceProvider()
       .describeUserPool({ UserPoolId: userpoolId })
+      .promise();
+  } catch (e) {
+    console.log(e);
+  }
+  return res;
+};
+
+const getLambdaFunction = async (functionName, region) => {
+  AWS.config.update({ region });
+  const lambda = new AWS.Lambda();
+  let res;
+  try {
+    res = await lambda
+      .getFunction({ FunctionName: functionName })
       .promise();
   } catch (e) {
     console.log(e);
@@ -41,4 +66,9 @@ const getBot = async (botName: string, region: string) => {
   return await service.getBot({ name: botName, versionOrAlias: '$LATEST' }).promise();
 };
 
-export { getUserPool, getUserPoolClients, getBot };
+const getFunction = async (functionName: string, region: string) => {
+  const service = new AWS.Lambda({ region });
+  return await service.getFunction({ FunctionName: functionName }).promise();
+};
+
+export { getDDBTable, checkIfBucketExists, getUserPool, getUserPoolClients, getBot, getLambdaFunction, getFunction };