+++ packages/amplify-provider-awscloudformation/lib/push-resources.js
@@ -12,6 +12,7 @@ const { transformGraphQLSchema } = require('./transform-graphql-schema');
 const { displayHelpfulURLs } = require('./display-helpful-urls');
 const { downloadAPIModels } = require('./download-api-models');
 const { loadResourceParameters } = require('../src/resourceParams');
+const { uploadAuthTriggerFiles } = require('./upload-auth-trigger-files');
 const archiver = require('../src/utils/archiver');
 
 const spinner = ora('Updating resources in the cloud. This may take a few minutes...');
@@ -62,6 +63,7 @@ async function run(context, category, resourceName) {
         );
       }
     })
+    .then(() => uploadAuthTriggerFiles(context, resourcesToBeCreated, resourcesToBeUpdated))
     .then(async () => {
       let {
         allResources,
@@ -245,10 +247,17 @@ function packageResources(context, resources) {
 
         const cfnMeta = context.amplify.readJsonFile(cfnFilePath);
 
-        cfnMeta.Resources.LambdaFunction.Properties.Code = {
-          S3Bucket: s3Bucket,
-          S3Key: s3Key,
-        };
+        if (cfnMeta.Resources.LambdaFunction.Type === 'AWS::Serverless::Function') {
+          cfnMeta.Resources.LambdaFunction.Properties.CodeUri = {
+            Bucket: s3Bucket,
+            Key: s3Key,
+          };
+        } else {
+          cfnMeta.Resources.LambdaFunction.Properties.Code = {
+            S3Bucket: s3Bucket,
+            S3Key: s3Key,
+          };
+        }
 
         const jsonString = JSON.stringify(cfnMeta, null, '\t');
         fs.writeFileSync(cfnFilePath, jsonString, 'utf8');