+++ packages/graphql-http-transformer/src/HttpTransformer.ts
@@ -97,7 +97,7 @@ export class HttpTransformer extends Transformer {
             const dataSourceID = HttpResourceIDs.HttpDataSourceID(baseURL)
             // only create one DataSource per base URL
             if (!ctx.getResource(dataSourceID)) {
-                ctx.addToStackMapping(HTTP_STACK_NAME, `^${dataSourceID}$`)
+                ctx.mapResourceToStack(HTTP_STACK_NAME, dataSourceID)
                 ctx.setResource(
                     dataSourceID,
                     this.resources.makeHttpDataSource(baseURL)
@@ -115,9 +115,9 @@ export class HttpTransformer extends Transformer {
         directive: DirectiveNode,
         ctx: TransformerContext
     ): void => {
-        ctx.addToStackMapping(
+        ctx.mapResourceToStack(
             HTTP_STACK_NAME,
-            `^${ResolverResourceIDs.ResolverResourceID(parent.name.value, field.name.value)}$`
+            ResolverResourceIDs.ResolverResourceID(parent.name.value, field.name.value)
         )
         const url: string = getDirectiveArgument(directive)("url")
         const baseURL: string = url.replace(HttpTransformer.urlRegex, '$1')