+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@1.5.4...amplify-category-analytics@1.5.5) (2019-07-10)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
+## [1.5.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@1.5.3...amplify-category-analytics@1.5.4) (2019-07-09)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [1.5.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@1.5.2...amplify-category-analytics@1.5.3) (2019-06-20)