+++ packages/amplify-category-interactions/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@1.6.0...amplify-category-interactions@1.6.1) (2019-07-10)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
+# [1.6.0](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@1.5.6...amplify-category-interactions@1.6.0) (2019-07-09)
+
+
+### Features
+
+* cognito + s3 + dyanmodb lambda trigger support ([#1783](https://github.com/aws-amplify/amplify-cli/issues/1783)) ([c6fc838](https://github.com/aws-amplify/amplify-cli/commit/c6fc838))
+
+
+
+
+
 ## [1.5.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@1.5.5...amplify-category-interactions@1.5.6) (2019-06-26)
 
 **Note:** Version bump only for package amplify-category-interactions